공통모듈로부터 restaurant 관심사로 기능 분리
