+++ src/features/restaurant/__mocks__/restaurantApi.js
@@ -1,5 +1,3 @@
 export const fetchRestaurants = jest.fn();
 
 export const fetchRestaurantById = jest.fn();
-
-export const authorize = jest.fn();
